requirements.txt
# Core
Django==5.2.1                       
asgiref==3.11.0
sqlparse==0.5.5
tzdata==2025.3

# REST Framework
djangorestframework==3.16.1
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
PyJWT==2.10.1

# Database
psycopg2-binary==2.9.11                 
dj-database-url==2.1.0

# Authentication & Security
python-dotenv==1.2.1
python-decouple==3.8

# CORS & Middleware
django-cors-headers==4.9.0              
whitenoise==6.11.0

# Filters & Extensions
django-filter==24.3
django-extensions==4.1

# Admin UI
django-admin-interface==0.32.0
django-colorfield==0.14.0

# Storage & Media
cloudinary==1.44.1
django-cloudinary-storage==0.3.0
django-storages==1.14.6
pillow==12.1.0

# HTTP & Networking
requests==2.32.5
urllib3==2.6.3
certifi==2026.1.4
charset-normalizer==3.4.4
idna==3.11

# Schema & Validation
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
referencing==0.37.0
rpds-py==0.30.0
attrs==25.4.0
inflection==0.5.1
uritemplate==4.2.0
packaging==26.0

# Utilities
python-slugify==8.0.4
text-unidecode==1.3
python-dateutil==2.9.0.post0
six==1.17.0
PyYAML==6.0.3
regex==2025.11.3
tqdm==4.67.1
click==8.3.1
typing_extensions==4.15.0
tomli==2.4.0
openpyxl==3.1.5
reportlab==4.2.5

# Formatting & Linting
djlint==1.36.4
jsbeautifier==1.15.4
cssbeautifier==1.15.4
EditorConfig==0.17.1
json5==0.13.0
pathspec==1.0.3
colorama==0.4.6

# Production
gunicorn==25.0.1

#migration files
**/migrations/*
!**/migrations/__init__.py
 frontend/